Adjust the height of the table to be adaptive
